auth-service/requirements.txt
asyncpg==0.20.1
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
databases==0.3.2
fastapi==0.58.1
h11==0.9.0
httptools==0.1.1
psycopg2-binary==2.8.5
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.4
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
